requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asyncio-dgram==2.2.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
discord==2.3.2
discord.py==2.4.0
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
lyricsgenius @ git+https://github.com/johnwmillr/LyricsGenius.git@ee6c7fbb826dcbd99d5a75c49fde3a55bee43014
mcrcon==0.7.0
mcstatus @ git+https://github.com/py-mine/mcstatus.git@fc3b96f614f44cbc74207022aa54c95c0bd7e44f
multidict==6.1.0
propcache==0.2.1
psycopg2==2.9.10
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
yarl==1.18.3